Handle `latest` and `current` as valid target-tli values
Updates the type of the `--target-tli` argument from `check_positive`
to new `check_tli` type which allows `current` and `latest` in addition
to positive integers.

`RecoveryExecutor._set_pitr_targets` is updated to resolve both
`current` and `latest` to the numeric values of those timelines so that
the existing behaviour around deciding whether to enable PITR by setting
`is_pitr` is maintained.

`Server.get_required_xlog_files` is also updated to resolve `current`
and `latest` to the numeric values so that the correct WALs can be
copied when `current` or `latest` are used.

This patch also removes redundant check in barman/recovery_executor.py:
The result of `target_tli != backup_info.timeline` has already been
checked when assigning to `d_tli` so there is no need to check it again
on L505.

Closes EnterpriseDB#305

Thanks to @kapouer for reporting this issue.