Increase coverage of test_get_required_xlog_files
Updates `test_get_required_xlog_files` so that the `target_tli` and
`target_time` options of `get_required_xlog_files` are covered. We
specifically make sure that we test the newly added "current" and
"latest" values for target_tli, as well as the older numeric and `None`
forms. We also verify that WALs are not returned after the value of
`target_time`.

Relates to #305.